pyproject.toml
[tool.poetry]
name = "ai4s-papers-service"
version = "0.1.1"
description = "AI4S papers management service API"
authors = ["felix5572@github.com"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.0"
django-ninja = "^1.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.0.0"
whitenoise = "^6.6.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-django = "^4.5.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3 